Enables pt-br l10n on Hugo
Creates an entry in the languages menu for Brazilian Portuguese with weight 5. Note that most systems and websites show languages in alphabetic order, or make use of analytics data to create a Suggested Languages section on the language selector. This is currently following a pattern of weighting established by other l10n efforts: add the latest translation to be submitted to the bottom of the list.